// frontend/src/components/radio/SpotifyPlayer/types.ts
export type AuthData = {
  access_token: string;
  refresh_token: string;
};